webrecorder/requirements.txt
bottle==0.12.13
bottle-cork
boto3
youtube_dl
itsdangerous
requests>=2.9.1
werkzeug
gevent-websocket
websocket-client
har2warc
fakeredis
